beta version

+++ src/main/java/com/matela/production/DTO/QuantiteActuelleAchatDTO.java
@@ -7,15 +7,17 @@ import lombok.Setter;
 
 import java.time.LocalDate;
 
-
 @Setter
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 public class QuantiteActuelleAchatDTO {
     private Long idAchat;
-    private Integer matierePremiereId;
+    private Long matierePremiereId;
     private Double quantiteActuelle;
     private Double prixAchat;
     private LocalDate dateAchat;
 }
+
+
+
